refactor: split control plane pairing acp resources
