+++ crates/daemon/src/control_plane_server.rs
@@ -62,7 +62,8 @@ use self::control::*;
 mod events;
 use self::events::*;
 mod resources;
-use self::resources::*;
+mod resources_pairing_acp;
+use self::resources_pairing_acp::*;
 mod resources_session_task_approval;
 use self::resources_session_task_approval::*;
 mod serve;
